+++ src/common/library.ml
@@ -84,7 +84,7 @@ module LibMap :
         List.fold_left Filename.concat root ks in
       let rec get (root, old_ks) ks map =
         if Logger.log_enabled () then
-          log_lib "get @[<hv>%S@ [%a]@ [%a]@ %a@]" root
+          log_lib "get @[<hv>\"%s\"@ [%a]@ [%a]@ %a@]" root
             (D.list D.string) old_ks (D.list D.string) ks (D.strmap pp) map;
         match ks with
         | []      -> concat root old_ks
@@ -154,10 +154,10 @@ let set_lib_root : string option -> unit = fun dir ->
           | 0 -> ()
           | _ ->
               fatal_msg "Library root cannot be set:@.";
-              fatal_no_pos "Command %S had a non-zero exit." cmd
+              fatal_no_pos "Command \"%s\" had a non-zero exit." cmd
           | exception Failure msg ->
               fatal_msg "Library root cannot be set:@.";
-              fatal_msg "Command %S failed:@." cmd;
+              fatal_msg "Command \"%s\" failed:@." cmd;
               fatal_no_pos "%s" msg
       end;
       (* Register the library root as part of the module mapping.
@@ -168,8 +168,7 @@ let set_lib_root : string option -> unit = fun dir ->
    the file name [fn] if [mn] is not already mapped and [fn] is a valid
    directory. In case of failure the program terminates and a graceful error
    message is displayed. *)
-let add_mapping : string * string -> unit = fun (mn, fn) ->
-  let mp = Path.of_string mn in
+let add_mapping : Path.t * string -> unit = fun (mp, fn) ->
   let fn =
     try Filename.realpath fn
     with Invalid_argument f -> fatal_no_pos "%s: No such file or directory" f
@@ -189,7 +188,7 @@ let file_of_path : Path.t -> string = fun mp ->
   try
     let fp = LibMap.get mp !lib_mappings in
     if Logger.log_enabled () then
-      log_lib "file_of_path @[<hv>%a@ %a@ = %S@]"
+      log_lib "file_of_path @[<hv>%a@ %a@ = \"%s\"@]"
         Path.pp mp LibMap.pp !lib_mappings fp;
     fp
   with LibMap.Root_not_set -> fatal_no_pos "Library root not set."
@@ -256,7 +255,7 @@ let path_of_file : (string -> string) -> string -> Path.t =
     String.sub base (len_fp + 1) (len_base - len_fp - 1)
   in
   let full_mp = mp @ List.map escape (String.split_on_char '/' rest) in
-  log_lib "path_of_file @[<hv>%S@ = %a@]" fname Path.pp full_mp;
+  log_lib "path_of_file @[<hv>\"%s\"@ = %a@]" fname Path.pp full_mp;
   full_mp
 
 (** [install_path fname] prefixes the filename [fname] by the path to the