+++ src/parsing/pretty.ml
@@ -13,8 +13,73 @@ open Syntax
 open Format
 open Core
 
+(** Keywords table. *)
+let keyword_table = Hashtbl.create 59
+
+let is_keyword : string -> bool = Hashtbl.mem keyword_table
+
+let _ = let open LpLexer in
+  List.iter (fun (s, k) -> Hashtbl.add keyword_table s k)
+    [ "abort", ABORT
+    ; "admit", ADMIT
+    ; "admitted", ADMITTED
+    ; "apply", APPLY
+    ; "as", AS
+    ; "assert", ASSERT
+    ; "assertnot", ASSERTNOT
+    ; "associative", ASSOCIATIVE
+    ; "assume", ASSUME
+    ; "begin", BEGIN
+    ; "builtin", BUILTIN
+    ; "commutative", COMMUTATIVE
+    ; "compute", COMPUTE
+    ; "constant", CONSTANT
+    ; "debug", DEBUG
+    ; "end", END
+    ; "fail", FAIL
+    ; "flag", FLAG
+    ; "focus", FOCUS
+    ; "generalize", GENERALIZE
+    ; "have", HAVE
+    ; "in", IN
+    ; "induction", INDUCTION
+    ; "inductive", INDUCTIVE
+    ; "infix", INFIX
+    ; "injective", INJECTIVE
+    ; "left", ASSOC(Pratter.Left)
+    ; "let", LET
+    ; "off", SWITCH(false)
+    ; "on", SWITCH(true)
+    ; "opaque", OPAQUE
+    ; "open", OPEN
+    ; "prefix", PREFIX
+    ; "print", PRINT
+    ; "private", PRIVATE
+    ; "proofterm", PROOFTERM
+    ; "protected", PROTECTED
+    ; "prover", PROVER
+    ; "prover_timeout", PROVER_TIMEOUT
+    ; "quantifier", QUANTIFIER
+    ; "refine", REFINE
+    ; "reflexivity", REFLEXIVITY
+    ; "require", REQUIRE
+    ; "rewrite", REWRITE
+    ; "right", ASSOC(Pratter.Right)
+    ; "rule", RULE
+    ; "sequential", SEQUENTIAL
+    ; "simplify", SIMPLIFY
+    ; "solve", SOLVE
+    ; "symbol", SYMBOL
+    ; "symmetry", SYMMETRY
+    ; "type", TYPE_QUERY
+    ; "TYPE", TYPE_TERM
+    ; "unif_rule", UNIF_RULE
+    ; "verbose", VERBOSE
+    ; "why3", WHY3
+    ; "with", WITH ]
+
 let raw_ident : string pp = fun ppf s ->
-  if LpLexer.is_keyword s then out ppf "{|%a|}" Print.pp_uid s
+  if is_keyword s then out ppf "{|%a|}" Print.pp_uid s
   else Print.pp_uid ppf s
 
 let ident : p_ident pp = fun ppf {elt;_} -> raw_ident ppf elt
@@ -185,10 +250,10 @@ let query : p_query pp = fun ppf { elt; _ } ->
   | P_query_debug(true ,s) -> out ppf "set debug \"+%s\"" s
   | P_query_debug(false,s) -> out ppf "set debug \"-%s\"" s
   | P_query_flag(s, b) ->
-      out ppf "set flag %S %s" s (if b then "on" else "off")
+      out ppf "set flag \"%s\" %s" s (if b then "on" else "off")
   | P_query_infer(t, _) -> out ppf "type %a" term t
   | P_query_normalize(t, _) -> out ppf "compute %a" term t
-  | P_query_prover s -> out ppf "set prover %S" s
+  | P_query_prover s -> out ppf "set prover \"%s\"" s
   | P_query_prover_timeout n -> out ppf "set prover_timeout %d" n
   | P_query_print None -> out ppf "print"
   | P_query_print(Some qid) -> out ppf "print %a" qident qid
@@ -218,7 +283,7 @@ let tactic : p_tactic pp = fun ppf { elt;  _ } ->
   | P_tac_solve -> out ppf "solve"
   | P_tac_sym -> out ppf "symmetry"
   | P_tac_why3 p ->
-      let prover ppf s = out ppf " %S" s in
+      let prover ppf s = out ppf " \"%s\"" s in
       out ppf "why3%a" (Option.pp prover) p
   end
 
@@ -257,7 +322,7 @@ let proof : (p_proof * p_proof_end) pp = fun ppf (p, pe) ->
 
 let command : p_command pp = fun ppf { elt; _ } ->
   begin match elt with
-  | P_builtin (s, qid) -> out ppf "@[builtin %S@ ≔ %a@]" s qident qid
+  | P_builtin (s, qid) -> out ppf "@[builtin \"%s\"@ ≔ %a@]" s qident qid
   | P_inductive (_, _, []) -> assert false (* not possible *)
   | P_inductive (ms, xs, i :: il) ->
       out ppf "@[<v>@[%a%a@]%a@,%a@,end@]"